tiny changes to experiment scripts

+++ experiments/astar_parser.py
@@ -47,7 +47,7 @@ class AStarParser(Parser):
         self.add_pattern("num_generated_until_last_f_layer", r"\[AStar\] Number of generated states until last f-layer: (\d+)", type=int)
         self.add_pattern("num_pruned_until_last_f_layer", r"\[AStar\] Number of pruned states until last f-layer: (\d+)", type=int)
         self.add_pattern("cost", r"\[AStar\] Plan cost: (.+)", type=float)
-        self.add_pattern("length", r"\[AStar\] Plan length: (.+)", type=float)
+        self.add_pattern("length", r"\[AStar\] Plan length: (.+)", type=int)
         self.add_pattern("exhausted", r"(\[AStar\] Exhausted!)", type=str)
         self.add_pattern("val_plan_invalid", r"(Plan invalid)", type=str)
