+++ experiments/brfs_parser.py
@@ -47,7 +47,7 @@ class BrfsParser(Parser):
         self.add_pattern("num_generated_until_last_g_layer", r"\[BrFS\] Number of generated states until last g-layer: (\d+)", type=int)
         self.add_pattern("num_pruned_until_last_g_layer", r"\[BrFS\] Number of pruned states until last g-layer: (\d+)", type=int)
         self.add_pattern("cost", r"\[BrFS\] Plan cost: (.+)", type=float)
-        self.add_pattern("length", r"\[BrFS\] Plan length: (.+)", type=float)
+        self.add_pattern("length", r"\[BrFS\] Plan length: (.+)", type=int)
         self.add_pattern("exhausted", r"(\[BrFS\] Exhausted!)", type=str)
         self.add_pattern("val_plan_invalid", r"(Plan invalid)", type=str)
